Restructuracion de carpetas. Creacion de pagina de Error y subrutas en la pagina juego. Correción de errores
