+++ src/pages/ErrorPage.js
@@ -1,0 +1,6 @@
+export default function ErrorPage(){
+    return(
+        <h1>Error 404</h1>
+    );
+
+}
